Add page with customization examples
