+++ apps/docs/.eleventy.js
@@ -1,5 +1,6 @@
 const { EleventyRenderPlugin } = require("@11ty/eleventy");
 const codeBlockDemo = require("./transformers/code-block-demo");
+const variablesPreview = require("./transformers/variables-preview");
 const markdownLibrary = require("./libraries/markdown-it");
 const CleanCSS = require("clean-css");
 
@@ -35,6 +36,7 @@ module.exports = function (eleventyConfig) {
   });
 
   eleventyConfig.addTransform('codeBlockDemo', codeBlockDemo);
+  eleventyConfig.addTransform('variablesPreview', variablesPreview);
 
   eleventyConfig.setUseGitIgnore(false);
 
@@ -42,11 +44,11 @@ module.exports = function (eleventyConfig) {
     return new CleanCSS({}).minify(code).styles;
   });
 
-  eleventyConfig.addFilter("publicComponentsFilter", function (components) {
+  eleventyConfig.addFilter("publicPageFilter", function (pages) {
     const isServing = process.argv.includes('--serve');
     return isServing
-		? components
-		: components.filter(component => component?.status !== 'underlying');
+		? pages
+		: pages.filter(page => page?.status !== 'underlying');
   });
 
   return {
