+++ apps/docs/_includes/navigation.njk
@@ -10,16 +10,17 @@
 	</vwc-nav-disclosure>
 
   <vwc-nav-disclosure label="Design System" open>
-    {%  for item in designs | sort(false, false, 'title') %}
+    {%  for item in designs | publicPageFilter | sort(false, false, 'title') %}
     <vwc-nav-item
 			href="{{ ('/designs/' + item.title) | url }}"
-			text="{{ item.title | replace("-", " ") | capitalize }}"
+			text="{{ item.title | replace("-", " ") | capitalize }}
+			{% if (item.status) == "underlying" %}🔒{% endif %}"
 			{% if ('/designs/' + item.title + '/') == page.url %}aria-current="page"{% endif %}></vwc-nav-item>
     {% endfor %}
   </vwc-nav-disclosure>
 
   <vwc-nav-disclosure label="Components" open>
-    {%  for item in components | publicComponentsFilter | sort(false, false, 'title') %}
+    {%  for item in components | publicPageFilter | sort(false, false, 'title') %}
     <vwc-nav-item
 			href="{{ ('/components/' + item.title) | url }}"
 			text="{{ item.title | replace("-", " ") | capitalize }}
